+++ cardano-api/src/Cardano/Api/Query.hs
@@ -124,6 +124,7 @@ import qualified Cardano.Chain.Update.Validation.Interface as Byron.Update
 
 import           Cardano.Ledger.Binary
 import qualified Cardano.Ledger.Binary.Plain as Plain
+import qualified Cardano.Ledger.Credential as Shelley
 import           Cardano.Ledger.Crypto (Crypto)
 import qualified Cardano.Ledger.Shelley.API as Shelley
 import qualified Cardano.Ledger.Shelley.Core as Core
@@ -140,8 +141,8 @@ import           Cardano.Api.Keys.Shelley
 import           Cardano.Api.Modes
 import           Cardano.Api.NetworkId
 import           Cardano.Api.ProtocolParameters
-import           Cardano.Api.TxBody
 import           Cardano.Api.Tx (eraProtVerLow)
+import           Cardano.Api.TxBody
 import           Cardano.Api.Value
 import           Data.Word (Word64)
 
@@ -289,6 +290,10 @@ data QueryInShelleyBasedEra era result where
     :: Maybe (Set PoolId)
     -> QueryInShelleyBasedEra era (SerialisedStakeSnapshots era)
 
+  QueryStakeDelegDeposits
+    :: Set StakeCredential
+    -> QueryInShelleyBasedEra era (Map StakeCredential Lovelace)
+
 instance NodeToClientVersionOf (QueryInShelleyBasedEra era result) where
   nodeToClientVersionOf QueryEpoch = NodeToClientV_9
   nodeToClientVersionOf QueryGenesisParameters = NodeToClientV_9
@@ -305,6 +310,7 @@ instance NodeToClientVersionOf (QueryInShelleyBasedEra era result) where
   nodeToClientVersionOf (QueryPoolState _) = NodeToClientV_14
   nodeToClientVersionOf (QueryPoolDistribution _) = NodeToClientV_14
   nodeToClientVersionOf (QueryStakeSnapshot _) = NodeToClientV_14
+  nodeToClientVersionOf (QueryStakeDelegDeposits _) = NodeToClientV_16
 
 deriving instance Show (QueryInShelleyBasedEra era result)
 
@@ -680,6 +686,11 @@ toConsensusQueryShelleyBased erainmode (QueryPoolDistribution poolIds) =
   where
     getPoolIds :: Set PoolId -> Set (Shelley.KeyHash Shelley.StakePool Consensus.StandardCrypto)
     getPoolIds = Set.map (\(StakePoolKeyHash kh) -> kh)
+toConsensusQueryShelleyBased erainmode (QueryStakeDelegDeposits stakeCreds) =
+    Some (consensusQueryInEraInMode erainmode (Consensus.GetStakeDelegDeposits stakeCreds'))
+  where
+    stakeCreds' :: Set (Shelley.StakeCredential Consensus.StandardCrypto)
+    stakeCreds' = Set.map toShelleyStakeCredential stakeCreds
 
 consensusQueryInEraInMode
   :: forall era mode erablock modeblock result result' xs.
@@ -936,6 +947,13 @@ fromConsensusQueryResultShelleyBased _ QueryStakeSnapshot{} q' r' =
   case q' of
     Consensus.GetCBOR Consensus.GetStakeSnapshots {} -> SerialisedStakeSnapshots r'
     _                                                -> fromConsensusQueryResultMismatch
+
+fromConsensusQueryResultShelleyBased _ QueryStakeDelegDeposits{} q' stakeCreds' =
+    case q' of
+      Consensus.GetStakeDelegDeposits{} -> Map.map fromShelleyLovelace
+                                         . Map.mapKeysMonotonic fromShelleyStakeCredential
+                                         $ stakeCreds'
+      _                                 -> fromConsensusQueryResultMismatch
 
 -- | This should /only/ happen if we messed up the mapping in 'toConsensusQuery'
 -- and 'fromConsensusQueryResult' so they are inconsistent with each other.